Replace -march with -mcpu for more hardware-specific optimisations
This allows optimisations specific to the CPU model as opposed to
the family. For example some versions of GCC can optimise for a
cortex-a15 cpu but don't define a armv7ve family. This change
improves performance by 3%-5% for the C, C++ and Fortran
implementations on Tegra K1 (Cortex-A15) with GCC 4.8.2-19ubuntu1.