Rename C-arm to C-fast and run it on X86
It turns out that this implementation is also the fastest one on
x86 (on an i5-3210M).
Note that there are two makefile targets, c-fast and c-fast-arm.
c-fast will build on both X86 and ARM, but it is essential to
run c-fast-arm on ARM, which is compiled in ARM mode (as opposed
to Thumb) for best performance.